+++ core/scaling.lua
@@ -2,8 +2,8 @@ local _, ns = ...
 local E, C = ns.E, ns.C
 
 -- Blizzard
-local GetCVar = _G.GetCVar
-local SetCVar = _G.SetCVar
+local GetCVar = C_CVar and C_CVar.GetCVar or _G.GetCVar
+local SetCVar = C_CVar and C_CVar.SetCVar or _G.SetCVar
 
 --------------------------------------------------
 -- UIScaling